UT-206: makes adjustments in response to PR comments.
refactor: adds return type to a few functions.
refactor: changes a few functions defined as constants with arrow functions, into function definitions.
chore: adds nullish coalescing to a number of 'useSelector(...)'.
chore: updates the order of import in App.tsx.
style: solidifies border-width and makes the color the thing that changes on selecting a file.
refactor: fixes a typo in a css class name.