UT-211: applies a few last changes before merge.
refactor: build a .exe file of the newest version.
refactor: changes close button of settings to say 'CLOSE SETTINGS' instead of 'DISCARD CHANGES', as the check if changes has been made has not been implemented for this branch, but will be during UT-210.
refactor: improves the logic return of 'isValidSwipe()' for readability.
refactor: improves log made if Cliptool fails to load a file for hiddenFiles.