UT-210: renames a few components and extract swipe logic to its own component
refactor: extract swipe component logic to its own component, meant to be used as a wrapper for what should be swipeable.
refactor: renames thumbnailUsingImage to image-thumbnail.
refactor: renamed thumbnailUsingText to text-thumbnail.
refactor: renamed thumbnail to be plural.
refactor: moves decision of used css from thumbnail-button to the places it is used instead.