UT-215: lowers brightness of tab color by using same color as active buttons.
refactor: moves definition for color used for active tabs or buttons to reusable variable.
refactor: applied new variable with active color instead of fixed color value.
refactor: adds a padding for buttons on settings to not look as cramped.
refactor: renames a css class name, to better match where it is used.